% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_mle.R
\name{asg}
\alias{asg}
\title{asg}
\usage{
asg(x, beta1, beta2, mu, h, sigma1, sigma2)
}
\arguments{
\item{x}{explanatory variable}

\item{beta1}{intercept in first half}

\item{beta2}{intercept in second half}

\item{mu}{peak week}

\item{h}{peak}

\item{sigma1}{variance in first half}

\item{sigma2}{variance in second half}
}
\value{
A function
}
\description{
A function for the ff of the ASG Distribution
}
\examples{
asg(x=c(1:33), beta1=-4, beta2=-1, mu=15, h=10, sigma1=10, sigma2=15)

}
