% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HorseshoeBFDAFC.R
\name{fusion_bfda}
\alias{fusion_bfda}
\title{horseshoe_bfda_fc}
\usage{
fusion_bfda(W, Z, E, model = "Indep", prior = "HSt", group = NULL,
  niter = 6000, nwarmup = niter/2, nchains = 3, thin = 1,
  inits = NULL)
}
\arguments{
\item{W}{matrix of response}

\item{E}{Eigen functions that correspond to W}

\item{model}{Which hierarchy structure to be used: 'Indep' (default), 'Group Indep', or 'Group Hier'}

\item{prior}{Which prior to use to specify model: 'HS'(Horseshoe), 'HSt'(Horseshoe with t), 'LASSO', 'FHS'(Finnish HS)}

\item{group}{Grouping variable to be included if model is Region or Season}

\item{niter}{number of interations to be run (default=4000)}

\item{nchains}{number of chains to be run (default=3)}

\item{thin}{when you want to thin (default=1)}

\item{inits}{Add specific initial values}

\item{nwarmiup}{number of iterations to be used as warmup (see link below)}
}
\value{
A MCMC object
}
\description{
Runs BFDA with horseshoe prior on betas for forecasting
}
\examples{

}
\seealso{
\url{http://mc-stan.org/users/documentation/}
}
