% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_mle.R
\name{normal_mle}
\alias{normal_mle}
\title{normal_mle}
\usage{
normal_mle(y, x, group, season, inits = NULL, ll = NULL,
  correction = TRUE)
}
\arguments{
\item{y}{response variable which follows binomial dist}

\item{x}{explanatory variable}

\item{group}{groups of response}

\item{season}{seasons in response}

\item{inits}{initial values for optim (defualt=NULL)}

\item{ll}{likelihood to be optimized (defualt=ASG)}
}
\value{
A function
}
\description{
A function which finds and formats the mle of ASG ff using normal data model
}
\examples{
asg_mle(y, x, group, seas)

}
